requirements.txt
discord.py
aiohttp
Flask
pytest
aiofiles

